fix package locations
