+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/jackson/modules/LedgerIndexSerializer.java
@@ -3,7 +3,7 @@ package org.xrpl.xrpl4j.model.jackson.modules;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
-import org.xrpl.xrpl4j.model.transactions.LedgerIndex;
+import org.xrpl.xrpl4j.model.client.specifiers.LedgerIndex;
 
 import java.io.IOException;
