+++ xrpl4j-model/src/test/java/org/xrpl/xrpl4j/model/client/specifiers/LedgerIndexShortcutTest.java
@@ -1,4 +1,4 @@
-package org.xrpl.xrpl4j.model.client.common;
+package org.xrpl.xrpl4j.model.client.specifiers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,7 +11,6 @@ import org.json.JSONException;
 import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
-import org.xrpl.xrpl4j.model.client.specifiers.LedgerIndexShortcut;
 import org.xrpl.xrpl4j.model.jackson.ObjectMapperFactory;
 
 public class LedgerIndexShortcutTest {
