+++ xrpl4j-model/src/test/java/org/xrpl/xrpl4j/model/client/specifiers/LedgerSpecifierTest.java
@@ -1,4 +1,4 @@
-package org.xrpl.xrpl4j.model.client.common;
+package org.xrpl.xrpl4j.model.client.specifiers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
@@ -16,11 +16,8 @@ import org.json.JSONException;
 import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
-import org.xrpl.xrpl4j.model.client.specifiers.LedgerIndexShortcut;
-import org.xrpl.xrpl4j.model.client.specifiers.LedgerSpecifier;
 import org.xrpl.xrpl4j.model.jackson.ObjectMapperFactory;
 import org.xrpl.xrpl4j.model.transactions.Hash256;
-import org.xrpl.xrpl4j.model.transactions.LedgerIndex;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
